requirements.txt
# Gold Breakout Bot - Dependencies
# Compatible with Python 3.11, 3.12, and 3.13

# Core Data Processing
pandas>=2.2.0
numpy>=1.26.0,<2.0.0
scipy>=1.11.0

# API Clients
twelvedata>=1.2.14
requests>=2.31.0

# Scheduling
schedule>=1.2.0

# Web Server (for Render keep-alive)
flask>=3.0.0
gunicorn>=21.2.0

# Machine Learning (for future Phase 3)
scikit-learn>=1.3.0
joblib>=1.3.0

# Environment Variables
python-dotenv>=1.0.0

# PostgreSQL (for Render persistent storage - set DATABASE_URL env var)
psycopg2-binary>=2.9.0